MOD: 处理dao执行sql前的convert value
